Made the split complete
